requirements.txt
Django==1.9.5
django-crispy-forms==1.6.1
django-filter==1.0.0
django-markdown-deux==1.0.5
django-pagedown==0.1.1
djangorestframework==3.4.0
djangorestframework-jwt==1.9.0
Markdown==2.6.7
markdown2==2.3.1
Pillow==3.4.1
PyJWT==1.4.2
